+++ src/main/java/com/devtraces/arterest/service/feed/FeedDeleteService.java
@@ -49,14 +49,14 @@ public class FeedDeleteService {
 			}
 		}*/
 		
-		// 삭제될 FeedHashtagMap 데이터 목록을 가져옴.
+		/*// 삭제될 FeedHashtagMap 데이터 목록을 가져옴.
 		List<FeedHashtagMap> feedHashtagMapList = feedHashtagMapRepository.findByFeed(feed);
 		
 		// FeedHashtagMap 테이블에서 관련 정보 모두 삭제.
 		feedHashtagMapRepository.deleteAllByFeedId(feedId);
 
 		// 사용되지 않는 Hashtag 삭제.
-		deleteNotUsingHashtag(feedHashtagMapList);
+		deleteNotUsingHashtag(feedHashtagMapList);*/
 
 		// 레디스에서 좋아요 개수 기록한 키-밸류 쌍 삭제.
 		likeNumberCacheRepository.deleteLikeNumberInfo(feedId);
@@ -83,16 +83,6 @@ public class FeedDeleteService {
 		);
 
 		// 피드 삭제.
-		feedRepository.deleteById(feedId);
-	}
-
-	void deleteNotUsingHashtag(List<FeedHashtagMap> feedHashtagMapList){
-		// 삭제된 FeedHashtagMap의 feedId에 매핑된 hashtagId가 더이상 FeedHashtagMap에 존재하지 않을 경우,
-		// 해당 hastagId를 Hashtag 테이블에서 삭제함.
-		for (FeedHashtagMap feedHashtagMap : feedHashtagMapList) {
-			if(!feedHashtagMapRepository.existsByHashtag(feedHashtagMap.getHashtag())){
-				hashtagRepository.deleteById(feedHashtagMap.getHashtag().getId());
-			}
-		}
+		// feedRepository.deleteById(feedId);
 	}
 }
