+++ src/main/java/com/devtraces/arterest/service/feed/application/FeedDeleteApplication.java
@@ -3,6 +3,7 @@ package com.devtraces.arterest.service.feed.application;
 import com.devtraces.arterest.common.exception.BaseException;
 import com.devtraces.arterest.model.feed.Feed;
 import com.devtraces.arterest.model.feed.FeedRepository;
+import com.devtraces.arterest.service.hashtag.HashtagService;
 import com.devtraces.arterest.service.s3.S3Service;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
@@ -15,6 +16,7 @@ public class FeedDeleteApplication {
 
     private final FeedRepository feedRepository;
     private final S3Service s3Service;
+    private final HashtagService hashtagService;
 
     // TODO 스프링 @Async를 사용해서 비동기 멀티 스레딩으로 처리하면 응답지연시간 최소화 가능.
     @Transactional
@@ -33,7 +35,11 @@ public class FeedDeleteApplication {
             }
         }
 
+        // 해시태그 관련 정보들을 전부 삭제한다.
+        hashtagService.deleteHashtagRelatedData(feed);
 
+        // 피드 삭제.
+        feedRepository.deleteById(feedId);
     }
 
 }
